resuelto, ver resultados

+++ IMRTsolver-masterPSO/src/solution/Particle.h
@@ -24,8 +24,8 @@ namespace imrt{
       int c2; //FACTOR GLOBAL, POR AHORA FIJO LUEGO POR PARAMETRO
       double fitness; //EVALUATION OF PCURRENT
       double bfitness; //EVALUATION OF PBEST;
-      int r1;
-      int r2;
+      double r1;
+      double r2;
       Plan PBest; //BEST GLOBAL VALUE
       Plan Pcurrent; //BEST PERSONAL VALUE
 
@@ -34,7 +34,7 @@ namespace imrt{
   	    Particle(Plan& P);
 
   	    //Function with the process for calculate the velocity of the particle using the stations
-        void Velocityupdate(Plan &GBest, int type, int w, int c1, int c2, int r1, int r2);
+        void Velocityupdate(Plan &GBest, int type, int w, int c1, int c2, double r1, double r2);
         //Function for calculate the position of the solution using PSO
         void calculateFitness();
         //Function for update Pbest
