+++ IMRTsolver-masterPSO/src/solution/Plan.h
@@ -82,7 +82,7 @@ public:
 	    return *s;
 	}
 
-	void updateVelocity(Plan *Bglobal, Plan *Pbest, Plan *current, int w, int c1, int c2, int r1, int r2);
+	void updateVelocity(Plan *Bglobal, Plan *Pbest, Plan *current, int w, int c1, int c2, double r1, double r2);
 
 	void updatePosition();
